2023/Finals/Misc/Queesy-Receipt/src/requirements.txt
qrcode==7.4.2
pillow